added animation directory
