+++ client/src/shapes/index.ts
@@ -11,9 +11,6 @@ import { star } from './shapes/star';
 import { image } from './shapes/image';
 import { circle } from './shapes/circle';
 import { initShapeCache } from './cacher';
-import type { SquareSchema } from '@shapes/square/types';
-import { shapeProps, type SchemaId, type Shape, type ShapeFactory, type WithId } from './types';
-import type { ObjectKeysToTuple } from '@utils/types';
 
 export const useOptimizedShapes = () => {
   const toOptimized = initShapeCache()
@@ -34,136 +31,3 @@ export const useOptimizedShapes = () => {
 }
 
 export type OptimizedShapes = ReturnType<typeof useOptimizedShapes>
-
-type AnimationDefinitionId = string
-
-type AnimationDefinition = {
-  /**
-   * a unique identifier for this definition
-   */
-  id: AnimationDefinitionId,
-  /**
-   * the duration of this animation (in milliseconds)
-   */
-  durationMs: number,
-}
-
-type ActiveAnimation = {
-  /**
-   * links the active animation to the animation definition
-   */
-  definitionId: AnimationDefinitionId,
-  /**
-   * unix timestamp when the animation started
-   */
-  startedAt: number,
-  /**
-   * number of times this animation will run before automatically stopping (can be fractional).
-   * set to `Infinity` to run animation indefinitely
-   */
-  runCount: number,
-}
-
-type ActiveAnimationWithDefinition = ActiveAnimation & AnimationDefinition
-
-export const useAnimatedShapes = () => {
-  /**
-   * a mapping between shapes (via ids) and the animations currently
-   * active/running on those shapes
-   */
-  const activeAnimations: Map<SchemaId, ActiveAnimation[]> = new Map()
-
-  /**
-   * returns the number of times the animation has completed as a float.
-   *
-   * @example
-   * const runs = getCurrentRunCount(animation) // 2.5
-   * // "animation" is half way through its 3rd run
-   */
-  const getCurrentRunCount = ({
-    durationMs,
-    startedAt,
-  }: Pick<ActiveAnimationWithDefinition, 'startedAt' | 'durationMs'>) => {
-    const timeElapsed = Date.now() - startedAt
-    return timeElapsed / durationMs
-  }
-
-  /**
-   * returns the current progress through the active animation cycle, as a value between 0 and 1.
-   * Useful for determining how far along the animation is within its current run.
-   *
-   * @example
-   * const progress = getAnimationProgress(animation); // 0.25
-   * // "animation" is 25% through its current cycle
-   */
-  const getAnimationProgress = ({
-    durationMs,
-    startedAt
-  }: Pick<ActiveAnimationWithDefinition, 'startedAt' | 'durationMs'>) => {
-    const timeElapsed = Date.now() - startedAt
-    return (timeElapsed % durationMs) / durationMs
-  }
-
-  const validPropsSet: ReadonlySet<keyof Shape> = new Set(shapeProps)
-
-  const squareWithId: ShapeFactory<WithId<SquareSchema>> = (schema) => {
-    return new Proxy(square(schema), {
-      get: (target, rawProp) => {
-        const prop = rawProp as keyof Shape
-
-        if (!validPropsSet.has(prop)) {
-          console.warn('invalid prop passed to animated shape')
-          return target[prop]
-        }
-
-        const animationsOnShape = activeAnimations.get(schema.id)
-
-        const notAnimating = !animationsOnShape || animationsOnShape.length === 0
-        if (notAnimating) return target[prop]
-
-        const animationDefs = animationsOnShape.map((a) => ({
-          def: getAnimationDefinition(a.definitionId),
-          ...a
-        }))
-
-        // cleanup expired animations
-        for (const animation of animationDefs) {
-          // const { timeAtStart, def: { durationMs } } = animation
-          // const timesRun = getTimesRun(animationDurationMs, Date.now() - timeBegun)
-          // const shouldRemove = timesRun >= 0.5
-          // if (shouldRemove) {
-          //   activeAnimations.delete(schema.id)
-          // }
-        }
-
-        // resolve the properties for the animated shape schema
-        const animatedSchema =
-
-
-        const prog = getAnimationProgress(timeBegun)
-
-        const modifiedSquare = square({
-          ...schema,
-          // at 0.5 needs to be at schema.size, at 1 needs to be at 0
-          size: prog < 0.5
-            ? schema.size * (prog * 2)
-            : schema.size * (2 * (1 - prog))
-        })
-
-        return modifiedSquare[prop]
-      }
-    })
-  }
-
-  return {
-    shapes: { square: squareWithId },
-    animation: {
-      start: (id: SchemaId) => {
-        activeAnimations.set(id, Date.now())
-      },
-      stop: (id: SchemaId) => {
-        activeAnimations.delete(id)
-      }
-    },
-  }
-}
